chore: Platform functions extension for EE (#19575)
Related to #15538

To enable adding a new action only for EE.

- Moved `PLATFORM_FUNCTIONS` and related types
- Moved `NavigationSagas` and added payload type to use in EE
