+++ app/client/src/actions/focusHistoryActions.ts
@@ -1,9 +1,18 @@
 import { FocusState } from "reducers/uiReducers/focusHistoryReducer";
-import { ReduxActionTypes } from "@appsmith/constants/ReduxActionConstants";
+import {
+  ReduxAction,
+  ReduxActionTypes,
+} from "@appsmith/constants/ReduxActionConstants";
 import { Location } from "history";
 import { AppsmithLocationState } from "utils/history";
 
-export const routeChanged = (location: Location<AppsmithLocationState>) => {
+export type RouteChangeActionPayload = {
+  location: Location<AppsmithLocationState>;
+};
+
+export const routeChanged = (
+  location: Location<AppsmithLocationState>,
+): ReduxAction<RouteChangeActionPayload> => {
   return {
     type: ReduxActionTypes.ROUTE_CHANGED,
     payload: { location },
